# vectorio: speed up polygon