--- a/locale/fr.po
+++ b/locale/fr.po
@@ -8,7 +8,7 @@ msgid ""
 msgstr ""
 "Project-Id-Version: 0.1\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2020-05-12 14:37+1000\n"
+"POT-Creation-Date: 2020-05-18 13:32-0700\n"
 "PO-Revision-Date: 2020-05-17 20:56+0000\n"
 "Last-Translator: Anonymous <noreply@weblate.org>\n"
 "Language-Team: French <https://later.unpythonic.net/projects/circuitpython/"
@@ -94,7 +94,7 @@ msgstr "La liste %q doit être une liste"
 msgid "%q must be >= 1"
 msgstr "%q doit être >=1"
 
-#: shared-bindings/vectorio/Polygon.c
+#: shared-module/vectorio/Polygon.c
 msgid "%q must be a tuple of length 2"
 msgstr "%q doit être un tuple de longueur 2"
 
@@ -534,7 +534,8 @@ msgstr "Impossible d'obtenir la taille du scalaire sans ambigüité"
 
 #: ports/stm/common-hal/pulseio/PWMOut.c
 msgid "Cannot vary frequency on a timer that is already in use"
-msgstr "Impossible de faire varier la fréquence sur une minuterie déjà utilisée"
+msgstr ""
+"Impossible de faire varier la fréquence sur une minuterie déjà utilisée"
 
 #: shared-module/bitbangio/SPI.c
 msgid "Cannot write without MOSI pin."
@@ -1293,12 +1294,16 @@ msgid ""
 "constructor"
 msgstr ""
 "Le brochage utilise %d octets par élément, ce qui consomme plus que le %d "
-"octets idéal. Si cela ne peut pas être évité, transmettez allow_inefficient ="
-" True au constructeur"
+"octets idéal. Si cela ne peut pas être évité, transmettez allow_inefficient "
+"= True au constructeur"
 
 #: py/builtinhelp.c
 msgid "Plus any modules on the filesystem\n"
 msgstr "Ainsi que tout autre module présent sur le système de fichiers\n"
+
+#: shared-module/vectorio/Polygon.c
+msgid "Polygon needs at least 3 points"
+msgstr ""
 
 #: shared-bindings/ps2io/Ps2.c
 msgid "Pop from an empty Ps2 buffer"
@@ -2063,7 +2068,8 @@ msgstr "ne peut pas réaliser un import relatif"
 
 #: extmod/ulab/code/ndarray.c
 msgid "cannot reshape array (incompatible input/output shape)"
-msgstr "ne peut pas remodeler le tableau (forme d'entrée / sortie incompatible)"
+msgstr ""
+"ne peut pas remodeler le tableau (forme d'entrée / sortie incompatible)"
 
 #: py/emitnative.c
 msgid "casting"
@@ -2190,10 +2196,6 @@ msgstr "division par zéro"
 #: py/objdeque.c
 msgid "empty"
 msgstr "vide"
-
-#: shared-bindings/vectorio/Polygon.c
-msgid "empty %q list"
-msgstr "liste %q vide"
 
 #: extmod/moduheapq.c extmod/modutimeq.c
 msgid "empty heap"
@@ -3226,8 +3228,8 @@ msgstr "'{' sans correspondance dans le format"
 msgid "unreadable attribute"
 msgstr "attribut illisible"
 
-#: shared-bindings/displayio/TileGrid.c shared-bindings/vectorio/Polygon.c
-#: shared-bindings/vectorio/VectorShape.c
+#: shared-bindings/displayio/TileGrid.c shared-bindings/vectorio/VectorShape.c
+#: shared-module/vectorio/Polygon.c
 msgid "unsupported %q type"
 msgstr "type %q non pris on charge"
 
@@ -3700,6 +3702,9 @@ msgstr "'step' nul"
 
 #~ msgid "either pos or kw args are allowed"
 #~ msgstr "soit 'pos', soit 'kw' est permis en argument"
+
+#~ msgid "empty %q list"
+#~ msgstr "liste %q vide"
 
 #~ msgid "expected a DigitalInOut"
 #~ msgstr "objet DigitalInOut attendu"