--- a/locale/sv.po
+++ b/locale/sv.po
@@ -7,7 +7,7 @@ msgid ""
 msgstr ""
 "Project-Id-Version: PACKAGE VERSION\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2020-05-12 14:37+1000\n"
+"POT-Creation-Date: 2020-05-18 13:32-0700\n"
 "PO-Revision-Date: 2020-05-17 20:56+0000\n"
 "Last-Translator: Anonymous <noreply@weblate.org>\n"
 "Language-Team: LANGUAGE <LL@li.org>\n"
@@ -89,7 +89,7 @@ msgstr "%q-listan måste vara en lista"
 msgid "%q must be >= 1"
 msgstr "%q måste vara >= 1"
 
-#: shared-bindings/vectorio/Polygon.c
+#: shared-module/vectorio/Polygon.c
 msgid "%q must be a tuple of length 2"
 msgstr "%q måste vara en tuple av längd 2"
 
@@ -1067,7 +1067,8 @@ msgstr "MicroPython fatalt fel."
 
 #: shared-bindings/audiobusio/PDMIn.c
 msgid "Microphone startup delay must be in range 0.0 to 1.0"
-msgstr "Startfördröjningen för mikrofonen måste vara i intervallet 0,0 till 1,0"
+msgstr ""
+"Startfördröjningen för mikrofonen måste vara i intervallet 0,0 till 1,0"
 
 #: ports/mimxrt10xx/common-hal/busio/SPI.c ports/stm/common-hal/busio/SPI.c
 msgid "Missing MISO or MOSI Pin"
@@ -1278,13 +1279,17 @@ msgid ""
 "bytes.  If this cannot be avoided, pass allow_inefficient=True to the "
 "constructor"
 msgstr ""
-"Pinout använder %d byte per element, vilket förbrukar mer än det idealiska %"
-"d byte.  Om detta inte kan undvikas, skicka allow_inefficient=True till "
+"Pinout använder %d byte per element, vilket förbrukar mer än det idealiska "
+"%d byte.  Om detta inte kan undvikas, skicka allow_inefficient=True till "
 "konstruktorn"
 
 #: py/builtinhelp.c
 msgid "Plus any modules on the filesystem\n"
 msgstr "Plus eventuella moduler i filsystemet\n"
+
+#: shared-module/vectorio/Polygon.c
+msgid "Polygon needs at least 3 points"
+msgstr ""
 
 #: shared-bindings/ps2io/Ps2.c
 msgid "Pop from an empty Ps2 buffer"
@@ -2162,10 +2167,6 @@ msgstr "division med noll"
 #: py/objdeque.c
 msgid "empty"
 msgstr "tom"
-
-#: shared-bindings/vectorio/Polygon.c
-msgid "empty %q list"
-msgstr "tom %q-lista"
 
 #: extmod/moduheapq.c extmod/modutimeq.c
 msgid "empty heap"
@@ -3188,8 +3189,8 @@ msgstr "omatchad '{' i format"
 msgid "unreadable attribute"
 msgstr "attribut kan inte läsas"
 
-#: shared-bindings/displayio/TileGrid.c shared-bindings/vectorio/Polygon.c
-#: shared-bindings/vectorio/VectorShape.c
+#: shared-bindings/displayio/TileGrid.c shared-bindings/vectorio/VectorShape.c
+#: shared-module/vectorio/Polygon.c
 msgid "unsupported %q type"
 msgstr "typ %q stöds inte"
 
@@ -3268,3 +3269,6 @@ msgstr "y-värde utanför intervall"
 #: py/objrange.c
 msgid "zero step"
 msgstr "noll steg"
+
+#~ msgid "empty %q list"
+#~ msgstr "tom %q-lista"