… shared-module/vectorio/Polygon.c