Takes polygon from 126k pixels per second fill to 240k pps fill
  on a reference 5 point star 50x66px polygon, updating both location and shape
  at 10hz.  Tested on an m4 express feather.

As a curiosity, the flat-out fill rate of a shape whose get_pixel is `return 0;`
  fills just shy of 375k pixels per second.

Baseline:
```
Polygon -> shape:{1020px,  8.1ms, 126603.0pps fill}  shape_pixels:{5859.4us total,  5.7us/px}
Polygon -> shape:{ 510px,  4.0ms, 126603.0pps fill}  shape_pixels:{2929.7us total,  5.7us/px}
```

This change:
```
Polygon -> shape:{1020px,  4.3ms, 238738.0pps fill}  shape_pixels:{2807.6us total,  2.8us/px}
Polygon -> shape:{ 510px,  2.1ms, 245759.0pps fill}  shape_pixels:{ 976.6us total,  1.9us/px}
```