@WarriorOfWire 2020-05-15
> Now needs a .po merge. (Allow maintainers to edit and I can fix it for you. ;-) )

Ah there's that silly button.  Fix away and thanks :-D