@WarriorOfWire 2020-05-16
This PR got too dirty; I've rebased the intended changes on master to keep the changelog clean.  Should be good to go.